+++ tests/schema_test.py
@@ -821,6 +821,7 @@ class VTDoc(DocumentElement):
     attr = Attribute('b')
     req_child = Child('c', VTElement, required=True)
     child = Child('d', VTElement)
+    multi = Child('e', VTElement, multiple=True)
 
     def __repr__(self):
         return 'VTDoc(req_attr={0!r}, req_child={1!r})'.format(
@@ -839,12 +840,32 @@ class VTDoc(DocumentElement):
                                req_child=TextElement(value='a'))),
      False, False),
     (VTDoc(req_attr='a', req_child=VTElement()), False, True),
+    (VTDoc(req_attr='a', req_child=VTElement(), multi=[]), False, True),
+    (VTDoc(req_attr='a', req_child=VTElement(), multi=[
+        VTElement()
+    ]), False, True),
+    (VTDoc(req_attr='a', req_child=VTElement(), multi=[
+        VTElement(req_attr='a', req_child=TextElement(value='a'))
+    ]), False, True),
     (VTDoc(req_attr='a', req_child=VTElement(req_attr='a')), False, True),
+    (VTDoc(req_attr='a', req_child=VTElement(req_attr='a'),
+           multi=[]), False, True),
     (VTDoc(req_attr='a',
-           req_child=VTElement(req_child=TextElement(value='a'))), False, True),
+           req_child=VTElement(req_child=TextElement(value='a')),
+           multi=[]), False, True),
     (VTDoc(req_attr='a',
            req_child=VTElement(req_attr='a',
-                               req_child=TextElement(value='a'))), True, True)
+                               req_child=TextElement(value='a')),
+           multi=[]), True, True),
+    (VTDoc(req_attr='a',
+           req_child=VTElement(req_attr='a',
+                               req_child=TextElement(value='a')),
+           multi=[VTElement()]), False, True),
+    (VTDoc(req_attr='a',
+           req_child=VTElement(req_attr='a',
+                               req_child=TextElement(value='a')),
+           multi=[VTElement(req_attr='a',
+                            req_child=TextElement(value='a'))]), True, True)
 ])
 def test_validate_recurse(element, recur_valid, valid):
     assert validate(element, recurse=True, raise_error=False) is recur_valid